Splitting helpers out of filesize
